@@ -69,12 +69,13 @@
   },
   "devDependencies": {
     "7zip": "0.0.6",
+    "@types/nedb": "^1.8.9",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
     "concurrently": "^3.6.1",
     "devtron": "^1.4.0",
-    "electron": "^7.1.10",
-    "electron-builder": "^20.38.5",
+    "electron": "^8.0.3",
+    "electron-builder": "^22.4.0",
     "electron-devtools-installer": "^2.2.4",
     "eslint-config-prettier": "^4.0.0",
     "eslint-plugin-prettier": "^3.0.1",
@@ -83,14 +84,12 @@
     "npm-run-all": "4.1.5",
     "prettier": "^1.16.4",
     "pretty-quick": "^1.10.0",
-    "react-scripts": "2.1.3",
+    "react-scripts": "^3.4.0",
     "rollup": "^1.32.0",
     "rollup-plugin-typescript": "^1.0.1",
     "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "wait-on": "^3.2.0",
-    "webpack": "4.42.0",
-    "webpack-cli": "^3.3.10"
+    "typescript": "^3.8.3",
+    "wait-on": "^3.2.0"
   },
   "browserslist": [
     ">0.2%",
